Combined ConnectTask and ResetTask. Deleted ResetTask.
